update error dumper
