+++ modules/error/view/403.html.twig
@@ -7,6 +7,6 @@
 {{ parent() }}
 {% endblock %}
 {% block content %}
-    {% trans %}Oops! Application encountered an internal error. If you're willing to help, please submit issue here: {% endtrans %}<a href="https://github.com/SlayerBirden/site/issues">https://github.com/SlayerBirden/site/issues</a>
+    {% trans %}Oops! You're trying to enter protected area without appropriate permissions.{% endtrans %}
 {% endblock %}
 {% block footer %}{% endblock %}
